- more debugging for distributed systems - added some simple tests - spd_denoiser updated

+++ examples/spd-denoiser/pow_table.hpp
@@ -8,7 +8,7 @@ public:
   pow_table(float exp) {
     table = (float *)malloc(256 * sizeof(float));
     for(int i=0; i<256; ++i)
-      table[i] = pow(i, exp);
+      table[i] = pow((float)i, exp);
   }
 
   ~pow_table() {
